Move types in common types file
