+++ src/containers/Outcomes/OutcomesTable.tsx
@@ -5,7 +5,7 @@ import usePageTitle from '../../utils/hooks/usePageTitle'
 import { useRouter } from '../../utils/hooks/useRouter'
 import { useUserState } from '../../contexts/UserState'
 import { useOutcomesTable } from '../../utils/hooks/useOutcomes'
-import { HeaderRow, OutcomeTableAPIQueries } from './types'
+import { HeaderRow, OutcomeTableAPIQueries } from '../../utils/types'
 import Markdown from '../../utils/helpers/semanticReactMarkdown'
 import { constructElement, formatCellText } from './helpers'
 import PaginationBar from '../../components/List/Pagination'
